Exercicios Python
Numeração de exercicios python de acordo com as listas presentes no site https://wiki.python.org.br/ListaDeExercicios

+++ Python/DevMartins/Estrutura Sequencial/02NumberWrittenWas.py
@@ -1,5 +1,5 @@
 # Entrada de dados pelo usuario
-x = int(input('Digite um numero: '))
+x = float(input('Digite um numero: '))
 
 # Exibição de resultado (numero informado pelo usuario)
 print(f'O numero informado foi {x}')
